refactor api routes as /api/v1
