--- ecom-backend/api/dependencies.py
@@ -1,9 +1,0 @@
-from core.database import SessionLocal
-from typing import Generator
-
-def get_db() -> Generator:
-    db = SessionLocal()
-    try:
-        yield db
-    finally:
-        db.close()
