+++ ecom-backend/core/database.py
@@ -8,3 +8,10 @@ DATABASE_URL = f"sqlite:///{settings.db_name}.db"
 engine = create_engine(DATABASE_URL)
 Base = declarative_base()
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
+
+def get_db() -> Generator:
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
