+++ ecom-backend/domains/carts/api/dependencies.py
@@ -1,4 +1,4 @@
-from api.dependencies import get_db
+from core.database import get_db
 from fastapi import Depends
 from domains.carts.repository import SQLAlchemyCartRepository
 from sqlalchemy.orm import Session
